Update pt 1
Some writing and reorganization, not done with putting stuff in, patience please! 😄 Moved the description of the variables to Part 1 since they're better placed in exploratory analysis (where you look at everything and examine it) instead of preparation (where you've already examined everything in the exploratory analysis and are changing things as needed).